+++ libs/designer/src/lib/core/state/operation/operationMetadataSlice.ts
@@ -78,9 +78,10 @@ export interface OperationMetadataState {
   operationInfo: Record<string, NodeOperation>;
   inputParameters: Record<string, NodeInputs>;
   outputParameters: Record<string, NodeOutputs>;
-  settings: Record<string, Settings>;
   dependencies: Record<string, NodeDependencies>;
   operationMetadata: Record<string, OperationMetadata>;
+  settings: Record<string, Settings>;
+  actionMetadata: Record<string, any>;
   staticResults: Record<string, NodeStaticResults>;
 }
 
@@ -91,6 +92,7 @@ const initialState: OperationMetadataState = {
   dependencies: {},
   settings: {},
   operationMetadata: {},
+  actionMetadata: {},
   staticResults: {},
 };
 
@@ -108,9 +110,10 @@ export interface NodeData {
   nodeInputs: NodeInputs;
   nodeOutputs: NodeOutputs;
   nodeDependencies: NodeDependencies;
-  settings?: Settings;
   operationMetadata: OperationMetadata;
   staticResults?: NodeStaticResults;
+  settings?: Settings;
+  actionMetadata?: Record<string, any>;
 }
 
 interface AddSettingsPayload {
@@ -155,11 +158,9 @@ export const operationMetadataSlice = createSlice({
     },
     initializeNodes: (state, action: PayloadAction<(NodeData | undefined)[]>) => {
       for (const nodeData of action.payload) {
-        if (!nodeData) {
-          return;
-        }
+        if (!nodeData) return;
 
-        const { id, nodeInputs, nodeOutputs, nodeDependencies, settings, operationMetadata, staticResults } = nodeData;
+        const { id, nodeInputs, nodeOutputs, nodeDependencies, settings, operationMetadata, actionMetadata, staticResults } = nodeData;
         state.inputParameters[id] = nodeInputs;
         state.outputParameters[id] = nodeOutputs;
         state.dependencies[id] = nodeDependencies;
@@ -171,6 +172,8 @@ export const operationMetadataSlice = createSlice({
         if (staticResults) {
           state.staticResults[id] = staticResults;
         }
+        if (actionMetadata) state.actionMetadata[id] = actionMetadata;
+        if (settings) state.settings[id] = settings;
       }
     },
     addDynamicInputs: (state, action: PayloadAction<AddDynamicInputsPayload>) => {
@@ -308,6 +311,13 @@ export const operationMetadataSlice = createSlice({
       const { id, nodeOutputs } = action.payload;
       if (state.outputParameters[id]) state.outputParameters[id] = nodeOutputs;
     },
+    updateActionMetadata: (state, action: PayloadAction<{ id: string; actionMetadata: Record<string, any> }>) => {
+      const { id, actionMetadata } = action.payload;
+      state.actionMetadata[id] = {
+        ...state.actionMetadata[id],
+        ...actionMetadata,
+      };
+    },
     deinitializeOperationInfo: (state, action: PayloadAction<{ id: string }>) => {
       const { id } = action.payload;
       delete state.operationInfo[id];
@@ -317,9 +327,10 @@ export const operationMetadataSlice = createSlice({
         delete state.inputParameters[id];
         delete state.outputParameters[id];
         delete state.dependencies[id];
-        delete state.settings[id];
         delete state.operationMetadata[id];
         delete state.staticResults[id];
+        delete state.settings[id];
+        delete state.actionMetadata[id];
       }
     },
   },
@@ -340,6 +351,7 @@ export const {
   updateParameterValidation,
   removeParameterValidationError,
   updateOutputs,
+  updateActionMetadata,
   deinitializeOperationInfo,
   deinitializeNodes,
 } = operationMetadataSlice.actions;