+++ libs/designer/src/lib/core/utils/parameters/helper.ts
@@ -26,6 +26,7 @@ import type { RootState } from '../../store';
 import { initializeArrayViewModel } from '../editors/array';
 import { getAllParentsForNode, getFirstParentOfType, getTriggerNodeId } from '../graph';
 import { getParentArrayKey, isForeachActionNameForLoopsource } from '../loops';
+import { isOneOf } from '../openapi/schema';
 import { loadDynamicOutputsInNode } from '../outputs';
 import { hasSecureOutputs } from '../setting';
 import { convertWorkflowParameterTypeToSwaggerType } from '../tokens';
@@ -959,6 +960,10 @@ export function loadParameterValuesFromDefault(inputParameters: Record<string, I
   }
 }
 
+export function loadParameterValuesArrayFromDefault(inputParameters: InputParameter[]): void {
+  for (const inputParameter of inputParameters) if (inputParameter.default !== undefined) inputParameter.value = inputParameter.default;
+}
+
 export function updateParameterWithValues(
   parameterKey: string,
   parameterValue: any,
@@ -1392,6 +1397,7 @@ export async function updateParameterAndDependencies(
   operationInfo: NodeOperation,
   connectionReference: ConnectionReference,
   nodeInputs: NodeInputs,
+  nodeMetadata: any,
   dependencies: NodeDependencies,
   variables: VariableDeclaration[],
   settings: Settings,
@@ -1452,6 +1458,7 @@ export async function updateParameterAndDependencies(
       connectionReference,
       dependenciesToUpdate,
       updateNodeInputsWithParameter(nodeInputs, parameterId, groupId, properties),
+      nodeMetadata,
       settings,
       variables,
       dispatch,
@@ -1506,6 +1513,7 @@ export async function updateDynamicDataInNode(
   connectionReference: ConnectionReference | undefined,
   dependencies: NodeDependencies,
   nodeInputs: NodeInputs,
+  nodeMetadata: any,
   settings: Settings,
   variables: VariableDeclaration[],
   dispatch: Dispatch,
@@ -1519,6 +1527,7 @@ export async function updateDynamicDataInNode(
     connectionReference,
     dependencies,
     nodeInputs,
+    nodeMetadata,
     settings,
     variables,
     dispatch,
@@ -1538,6 +1547,7 @@ export async function updateDynamicDataInNode(
           operationInfo,
           connectionReference,
           nodeInputs,
+          nodeMetadata,
           dependencies,
           false /* showErrorWhenNotReady */,
           dispatch
@@ -1554,6 +1564,7 @@ async function loadDynamicData(
   connectionReference: ConnectionReference | undefined,
   dependencies: NodeDependencies,
   nodeInputs: NodeInputs,
+  nodeMetadata: any,
   settings: Settings,
   variables: VariableDeclaration[],
   dispatch: Dispatch,
@@ -1561,7 +1572,17 @@ async function loadDynamicData(
   operationDefinition?: any
 ): Promise<void> {
   if (Object.keys(dependencies.outputs).length) {
-    loadDynamicOutputsInNode(nodeId, isTrigger, operationInfo, connectionReference, dependencies.outputs, nodeInputs, settings, dispatch);
+    loadDynamicOutputsInNode(
+      nodeId,
+      isTrigger,
+      operationInfo,
+      connectionReference,
+      dependencies.outputs,
+      nodeInputs,
+      nodeMetadata,
+      settings,
+      dispatch
+    );
   }
 
   if (Object.keys(dependencies.inputs).length) {
@@ -1571,6 +1592,7 @@ async function loadDynamicData(
       operationInfo,
       connectionReference,
       nodeInputs,
+      nodeMetadata,
       variables,
       dispatch,
       rootState,
@@ -1585,6 +1607,7 @@ async function loadDynamicContentForInputsInNode(
   operationInfo: NodeOperation,
   connectionReference: ConnectionReference | undefined,
   allInputs: NodeInputs,
+  nodeMetadata: any,
   variables: VariableDeclaration[],
   dispatch: Dispatch,
   rootState: RootState,
@@ -1596,7 +1619,7 @@ async function loadDynamicContentForInputsInNode(
       dispatch(clearDynamicInputs(nodeId));
 
       if (isDynamicDataReadyToLoad(info)) {
-        const inputSchema = await getDynamicSchema(info, allInputs, operationInfo, connectionReference, variables);
+        const inputSchema = await getDynamicSchema(info, allInputs, nodeMetadata, operationInfo, connectionReference, variables);
         const allInputParameters = getAllInputParameters(allInputs);
         const allInputKeys = allInputParameters.map((param) => param.parameterKey);
         const schemaInputs = inputSchema
@@ -1627,6 +1650,7 @@ export async function loadDynamicValuesForParameter(
   operationInfo: NodeOperation,
   connectionReference: ConnectionReference | undefined,
   nodeInputs: NodeInputs,
+  nodeMetadata: any,
   dependencies: NodeDependencies,
   showErrorWhenNotReady: boolean,
   dispatch: Dispatch,
@@ -1655,7 +1679,14 @@ export async function loadDynamicValuesForParameter(
       );
 
       try {
-        const dynamicValues = await getDynamicValues(dependencyInfo, nodeInputs, operationInfo, connectionReference, idReplacements);
+        const dynamicValues = await getDynamicValues(
+          dependencyInfo,
+          nodeInputs,
+          nodeMetadata,
+          operationInfo,
+          connectionReference,
+          idReplacements
+        );
 
         dispatch(
           updateNodeParameters({
@@ -1927,7 +1958,6 @@ function swapInputsValueIfNeeded(inputsValue: any, manifest: OperationManifest) 
     deleteObjectProperty(finalValue, target);
     finalValue = !source.length ? { ...finalValue, ...value } : safeSetObjectPropertyValue(finalValue, source, value);
   }
-
   return finalValue;
 }
 
@@ -2399,11 +2429,7 @@ export function parameterValueToString(
     return castParameterValueToString(value, parameterFormat, parameterType);
   }
 
-  if (
-    parameterType === constants.SWAGGER.TYPE.OBJECT ||
-    parameterType === constants.SWAGGER.TYPE.ARRAY ||
-    (parameter.schema && parameter.schema['oneOf'])
-  ) {
+  if (parameterType === constants.SWAGGER.TYPE.OBJECT || parameterType === constants.SWAGGER.TYPE.ARRAY || isOneOf(parameter.schema)) {
     return parameterValueToJSONString(value, /* applyCasting */ !parameterSuppressesCasting);
   }